refact: error hadnling in process_philo in bonus

+++ philo_bonus/inc/philo_bonus.h
@@ -95,7 +95,7 @@ void		*mgo_calloc(size_t count, size_t size);
 void		mgo_putstr_fd(char *s, int fd);
 size_t		mgo_strlen(const char *s);
 int			error_with_msg(char *msg);
-long int	set_time_ms(long int *dest);
+int			set_time_ms(long int *dest);
 int			sleep_shortly_ms(long int ms_sleep);
 int			print_philo_died(t_philo *philo, long int ms_now);
 int			print_philo_status(t_philo *philo, char *status);
